fix config.h location and incorrectly checked in copy, fix ALIVE_FATAL deps hack fix to try to solve gcc 10.X link issue
