+++ Source/AliveLibAE/bmp.cpp
@@ -3,6 +3,7 @@
 #include "Function.hpp"
 #include "DDraw.hpp"
 #include "Error.hpp"
+#include "Sys_common.hpp"
 
 ALIVE_VAR(1, 0xBC0BB4, unsigned __int8, gVGA_force_sys_memory_surfaces_BC0BB4, 0);
