--- Source/AliveLibAE/config.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#define DEVELOPER_MODE 1
-#define BEHAVIOUR_CHANGE_FORCE_WINDOW_MODE 1
-#define BEHAVIOUR_CHANGE_SUB_DATA_FOLDERS 1
-#define ORIGINAL_PS1_BEHAVIOR 1
-#define FORCE_DDCHEAT 1
-#define LCD_PS1_SPEED 0
-#define XINPUT_SUPPORT 0
-#define RENDER_TEST 0
-#define USE_SDL2 1
-#define USE_SDL2_SOUND 1
-#define USE_SDL2_IO 1
-/* #undef BUILD_NUMBER */
