+++ Source/AliveLibCommon/FunctionFwd.hpp
@@ -26,8 +26,6 @@ bool RunningAsInjectedDll();
 #define ALIVE_ASSERT_SIZEOF_ALWAYS(structureName, expectedSize) static_assert(sizeof(structureName) == expectedSize, "sizeof(" #structureName ") must be " #expectedSize)
 
 
-[[noreturn]] void ALIVE_FATAL(const char* errMsg);
-
 #define STATIC_EQUALS(src, dst) static_assert(src == dst, "Not equal!");
 #define ALIVE_COUNTOF(x) (sizeof(x)/sizeof(*(x)))
 #define BYTEn(x, n)   (*((BYTE*)&(x)+n))
