+++ options.cmake
@@ -12,4 +12,4 @@ option(USE_SDL2_SOUND "Use SDL2 for audio." ON)
 option(USE_SDL2_IO "Use SDL2 for all File/Stream IO." ON)
 option(ORIGINAL_PS1_BEHAVIOR "Fixes bugs in the PSX Emu layer / Gameplay to match PS1 version of the game." ON)
 
-CONFIGURE_FILE(${CMAKE_CURRENT_SOURCE_DIR}/Source/config.h.in ${CMAKE_CURRENT_SOURCE_DIR}/Source/AliveLib/config.h)
+CONFIGURE_FILE(${CMAKE_CURRENT_SOURCE_DIR}/Source/config.h.in ${CMAKE_CURRENT_SOURCE_DIR}/Source/AliveLibCommon/config.h)
